🚚 rename: type 및 ErrorPage 가져오기 업데이트 #46
